correção de casas decimais do peso do animal

+++ src/main/java/com/binno/dominio/module/animal/api/dto/PesoDto.java
@@ -1,6 +1,5 @@
 package com.binno.dominio.module.animal.api.dto;
 
-import com.binno.dominio.module.animal.model.Animal;
 import com.binno.dominio.module.animal.model.PesoAnimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -16,7 +15,7 @@ import java.util.stream.Collectors;
 public final class PesoDto {
     private final Integer id;
     private final LocalDate dataPesagem;
-    private final float peso;
+    private final Double peso;
 
     public static List<PesoDto> listToDto(List<PesoAnimal> pesoAnimal) {
         return pesoAnimal.stream()
